fetch/requirements.txt
apache-beam==2.2.0
google-cloud==0.32.0
tensorflow==1.6.0rc1
astral==1.5
geopy==1.11.0
numpy==1.14.0
pandas==0.22.0
scikit-learn==0.19.1